% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{alphabet}
\alias{alphabet}
\title{alphabet polygon data}
\usage{
data(alphabet)
}
\description{
polygons of all upper case letters
}
\details{
alphabet polygon data is based on upper case letters from Helvetica.
}

