% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alphabet.r
\name{createPolygons}
\alias{createPolygons}
\title{Create a data set of polygons for a set of letters from a specified fontfamily}
\usage{
createPolygons(letters, fontfamily, fontsize = 400, dim = c(720, 720))
}
\arguments{
\item{letters}{set of characters for which polygons are to be created}

\item{fontfamily}{character describing the name of a font - use fonts() from package extrafont to check on available fonts}

\item{fontsize}{size of letter to be created - larger means larger resolution}

\item{dim}{dimensions of the box in which the created letter is supposed to fit}
}
\value{
data set compatible to work with geom_logo
}
\description{
Create a data set of polygons for a set of letters from a specified fontfamily
}
\examples{
new_alphabet <- createPolygons(c(letters, LETTERS, 0:9), fontfamily="Garamond")
# check that all letters and digits are nicely shaped:
qplot(x,y, geom="polygon", data=new_alphabet, facets=~region)
}

