\name{point_line_dist}
\alias{point_line_dist}
\title{Distamce between points and line}
\usage{
  point_line_dist(px, py, lx_1, ly_1, lx_2, ly_2)
}
\arguments{
  \item{px}{x coordinate of point}

  \item{py}{y coordinate of point}

  \item{lx_1,}{}

  \item{ly_1,}{}

  \item{lx_2,}{}

  \item{ly_2}{x and y coordinates of points spanning a
  line}
}
\description{
  From
  http://mathworld.wolfram.com/Point-LineDistance2-Dimensional.html
}

