\docType{data}
\name{sequences}
\alias{sequences}
\title{peptide sequence data}
\description{
  available through biovis redesign contest 2013, see
  http://www.biovis.net/year/2013/info/redesign-contest
  published in Wong, B. Nat Methods 7, 889 (2011)
}
\details{
  Sequence data
}
\keyword{datasets}

