% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logos.r
\name{splitSequence}
\alias{splitSequence}
\title{Reshape data set according to elements in sequences}
\usage{
splitSequence(dframe, sequences)
}
\arguments{
\item{dframe}{data frame of peptide (or any other) sequences and some treatment factors}

\item{sequences}{column containing the character vector of (peptide) sequence}
}
\description{
prepare data set for plotting in a logo
}
\examples{
data(sequences)
dm2 <- splitSequence(sequences, peptide)
}
