\docType{data}
\name{StatLogo}
\alias{StatLogo}
\title{Helper object}
\format{proto object 
 $ calculate_groups:function (., data, na.rm = FALSE, width = width, ...)   
  ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 214 23 230 3 23 3 214 230 
  .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x122592178>  
 $ objname         : chr "logo" 
 $ default_geom    :function (.)   
  ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 240 19 240 38 19 38 240 240 
  .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x122592178>  
 $ calculate       :function (., data, scales, binwidth = NULL, origin = NULL, breaks = NULL, width = 0.9, na.rm = FALSE,  
    ...)   
  ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 232 16 238 3 16 3 232 238 
  .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x122592178>  
 $ required_aes    : chr [1:4] "x" "y" "group" "label" 
 parent: proto object 
 .. parent: proto object}
\usage{
  StatLogo
}
\value{
  proto object
}
\description{
  shouldn't be used by the user, but has to be exported.
}
\keyword{datasets}

