% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{alphabet}
\alias{alphabet}
\alias{alphabet_comic}
\alias{alphabet_xkcd}
\alias{alphabet_braille}
\title{alphabet polygon data}
\usage{
data(alphabet)

data(alphabet_comic)

data(alphabet_xkcd)

data(alphabet_braille)
}
\description{
polygons of upper and lower case letters and digits 0-9
}
\details{
alphabet polygon data

The alphabet polygon data is based on the font Helvetica. Upper and lower case letters are included as well as digits.
Other alphabet data are `alphabet_comic` (based on Comic Sans MS), `alphabet_xkcd` (based on the xkcd font),  and `alphabet_braille` (based on Swell Braille)
}
