% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{alphabet}
\alias{alphabet}
\title{alphabet polygon data}
\usage{
data(alphabet)
}
\description{
polygons of upper and lower case letters and digits 0-9
}
\details{
alphabet polygon data

The alphabet polygon data is based on the font Helvetica. Upper and lower case letters are included as well as digits.
}

