% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text.r
\name{determineOrder}
\alias{determineOrder}
\title{Determine order in which to pass through a set of points}
\usage{
determineOrder(x, y)
}
\arguments{
\item{x}{x coordinates}

\item{y}{y coordinates}
}
\value{
vector of indices for ordered traversion along border
}
\description{
Greedy algorithm to connect points, with the idea that a point is connected by a line with the two points closest to each, that haven't yet been connected into the shape.
Results depend on the starting point.
}

