% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/logos.r
\name{geom_logo}
\alias{geom_logo}
\title{Sequence logo plots.}
\usage{
geom_logo(mapping = NULL, data = NULL, stat = "logo",
  position = "identity", show.legend = NA, inherit.aes = TRUE,
  width = 0.9, alpha = 0.25, ...)
}
\arguments{
\item{mapping}{The aesthetic mapping, usually constructed with aes or aes_string. Only needs to be set at the layer level if you are overriding the plot defaults.}

\item{data}{A layer specific dataset - only needed if you want to override the plot defaults,}

\item{stat}{The statistical transformation to use on the data for this layer,}

\item{position}{The position adjustment to use for overlappling points on this layer,}

\item{show.legend}{Whether to show the legend or not}

\item{inherit.aes}{Whether to inherit the aes or not}

\item{width}{maximum width of the letters, defaults to 0.9,}

\item{alpha}{amount of alpha blending used for putting letters on top of rectangle, defaults to 0.25,}

\item{...}{other arguments passed on to layer. This can include aesthetics whose values you want to set, not map. See layer for more details.}
}
\description{
Sequence logo plots.
}
\examples{
\donttest{
data(sequences)
dm2 <- splitSequence(sequences, "peptide")
dm3 <- calcInformation(dm2, pos="position", elems="element", k=21)
#library(biovizBase)
#cols <- getBioColor(type="AA_ALPHABET")
library(RColorBrewer)
cols <- brewer.pal(10,"Paired")
data(aacids)
dm3b <- merge(dm3, aacids, by.x="element", by.y="AA", all.x=T)
ggplot(dm3b, aes(x=position, y=bits, group=element,
     label=element, fill=interaction(Polarity, Water))) +
     geom_logo(alpha=0.6) + scale_fill_brewer(palette="Paired")
dm4 <- calcInformation(dm2, pos="position", elems="element", trt="class", k=21)
cols2 <- alpha(cols, 0.8)
dm5 <- merge(dm4, aacids, by.x="element", by.y="AA", all.x=T)
ggplot(dm4, aes(x=class, y=bits, group=element,
     label=element, fill=element)) +
     geom_logo(alpha=0.8) + facet_wrap(~position, ncol=18)
ggplot(dm4, aes(x=position, y=bits, group=element, label=element, fill=element)) +
     geom_logo(alpha=0.8) + #scale_fill_manual(values=alpha(cols, 0.8)) +
     facet_wrap(~class, ncol=1) + theme_bw()
ggplot(dm5, aes(x=class, y=bits, group=element,
     label=element, fill=interaction(Polarity, Water))) +
     geom_logo(alpha=0.8) + scale_fill_brewer("Amino-acids properties", palette="Paired") +
     facet_wrap(~position, ncol=18) + theme(legend.position="bottom") + xlab("")
ggplot(dm5, aes(x=class, y=bits, group=element,
     label=element, fill=interaction(Water, Polarity))) +
     geom_logo(alpha=0.8) + scale_fill_brewer("Amino acids properties", palette="Paired") +
     facet_wrap(~position, ncol=18) + theme_bw() + theme(legend.position="bottom") +
     xlab("") + ylab("Shannon information in bits")
}
}

