% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logos.r
\name{logo}
\alias{logo}
\title{Logo plot}
\usage{
logo(sequences)
}
\arguments{
\item{sequences}{vector of text sequences, for which consensus logo is to be shown}
}
\value{
ggplot2 object for simple sequence
}
\description{
Simple logo plot of sequences. For more complicated sequence logos, such as with treatment comparisons or subsets see geom_logo.
}
\examples{
data(sequences)
library(ggplot2)
library(RColorBrewer)
cols <- rep(brewer.pal(12, name="Paired"),22)
logo(sequences$peptide) + aes(fill=element) + scale_fill_manual(values=cols)
}

