% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thin.r
\name{point_line_dist}
\alias{point_line_dist}
\title{Distance between point and line}
\usage{
point_line_dist(px, py, lx_1, ly_1, lx_2, ly_2)
}
\arguments{
\item{px}{x coordinate of  point outside the}

\item{py}{y coordinate of  point}

\item{lx_1, }{x  coordinate of 1st point spanning a line}

\item{ly_1, }{y  coordinate of 1st point spanning a line}

\item{lx_2, }{x  coordinate of 2nd point spanning a line}

\item{ly_2, }{y  coordinate of 2nd point spanning a line}
}
\description{
Compute distance between point given as (px, py) and line spanned by points (lx1, ly1) and (lx2, ly2).
From http://mathworld.wolfram.com/Point-LineDistance2-Dimensional.html
}

