% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{sequences}
\alias{sequences}
\title{peptide sequence data}
\usage{
data(sequences)
}
\description{
sequences showing the amino acid usage in the adenylate kinase lid available through biovis redesign contest 2013, see \url{http://biovis.net/year/2013/info/redesign-contest}
published in Wong, B. Nat Methods 7, 889 (2011)
}
\details{
Sequence data
}
