% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/logos.r
\name{stat_logo}
\alias{stat_logo}
\title{Calculation of all pieces necessary to plot a logo sequence plot}
\usage{
stat_logo(mapping = NULL, data = NULL, geom = "logo",
  position = "identity", width = 0.9, ...)
}
\arguments{
\item{mapping}{The aesthetic mapping, usually constructed with aes or aes_string. Only needs to be set at the layer level if you are overriding the plot defaults.}

\item{data}{A layer specific dataset - only needed if you want to override the plot defaults,}

\item{geom}{The geometric object to use display the data,}

\item{position}{The position adjustment to use for overlappling points on this layer,}

\item{width}{maximum width of the letters, defaults to 0.9,}

\item{...}{other arguments passed on to layer. This can include aesthetics whose values you want to set, not map. See layer for more details.}
}
\value{
A proto object
}
\description{
Calculation of all pieces necessary to plot a logo sequence plot
}
\examples{
# See geom_logo for examples
# Generate data
data(sequences)
library(ggplot2)
dm2 <- splitSequence(sequences, "peptide")
dm3 <- calcInformation(dm2, pos="position", elems="element", k=21)
ggplot(dm3, aes(x=position, y=bits, label=element, group=interaction(position, element))) +
     geom_logo()
}

