% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logos.r
\docType{data}
\name{StatLogo}
\alias{StatLogo}
\alias{stat_logo}
\title{Calculation of all pieces necessary to plot a logo sequence plot}
\format{An object of class \code{StatLogo} (inherits from \code{Stat}, \code{ggproto}, \code{gg}) of length 4.}
\usage{
StatLogo

stat_logo(mapping = NULL, data = NULL, geom = "logo",
  position = "logo", show.legend = NA, inherit.aes = TRUE,
  width = 0.9, na.rm = TRUE, ...)
}
\arguments{
\item{mapping}{The aesthetic mapping, usually constructed with aes or aes_string. Only needs to be set at the layer level if you are overriding the plot defaults.}

\item{data}{A layer specific dataset - only needed if you want to override the plot defaults,}

\item{geom}{The geometric object to use display the data,}

\item{position}{The position adjustment to use for overlappling points on this layer,}

\item{show.legend}{Whether to show the legend or not}

\item{inherit.aes}{Whether to inherit the aes or not}

\item{width}{maximum width of the letters, defaults to 0.9,}

\item{na.rm}{Whether to remove NAs or not}

\item{...}{other arguments passed on to layer. This can include aesthetics whose values you want to set, not map. See layer for more details.}
}
\value{
proto object

A proto object
}
\description{
Calculation of all pieces necessary to plot a logo sequence plot
}
\examples{
# See geom_logo for examples
# Generate data
data(sequences)
library(ggplot2)

ggplot(data = ggfortify(sequences, "peptide")) + 
  geom_logo(aes(x=position, y=bits, label=element, 
                group=interaction(position, element)), 
            alpha=0.5)
}
\keyword{datasets}
