\docType{data}
\name{StatLogo}
\alias{StatLogo}
\title{Helper object}
\format{proto object 
 $ calculate_groups:function (., data, na.rm = FALSE, width = width, ...)   
  ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 213 23 229 3 23 3 213 229 
  .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x104e52910>  
 $ objname         : chr "logo" 
 $ default_geom    :function (.)   
  ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 239 19 239 38 19 38 239 239 
  .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x104e52910>  
 $ calculate       :function (., data, scales, binwidth = NULL, origin = NULL, breaks = NULL, width = 0.9, na.rm = FALSE,  
    ...)   
  ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 231 16 237 3 16 3 231 237 
  .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x104e52910>  
 $ required_aes    : chr [1:4] "x" "y" "group" "label" 
 parent: proto object 
 .. parent: proto object}
\usage{
  StatLogo
}
\value{
  proto object
}
\description{
  shouldn't be used by the user, but has to be exported.
}
\keyword{datasets}

