% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/logos.r
\docType{data}
\name{StatLogo}
\alias{StatLogo}
\title{Helper object}
\format{\preformatted{Classes 'proto', 'environment' <environment: 0x1036c2510> 
}}
\usage{
StatLogo
}
\value{
proto object
}
\description{
shouldn't be used by the user, but has to be exported.
}
\keyword{datasets}

