\docType{data}
\name{StatLogo}
\alias{StatLogo}
\title{Helper object}
\format{proto object 
 $ calculate_groups:function (., data, na.rm = FALSE, width = width, ...)   
  ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 215 23 231 3 23 3 215 231 
  .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x12b7d97c0>  
 $ objname         : chr "logo" 
 $ default_geom    :function (.)   
  ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 241 19 241 38 19 38 241 241 
  .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x12b7d97c0>  
 $ calculate       :function (., data, scales, binwidth = NULL, origin = NULL, breaks = NULL, width = 0.9, na.rm = FALSE,  
    ...)   
  ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 233 16 239 3 16 3 233 239 
  .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x12b7d97c0>  
 $ required_aes    : chr [1:4] "x" "y" "group" "label" 
 parent: proto object 
 .. parent: proto object}
\usage{
  StatLogo
}
\value{
  proto object
}
\description{
  shouldn't be used by the user, but has to be exported.
}
\keyword{datasets}

