% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/logos.r
\docType{data}
\name{StatLogo}
\alias{StatLogo}
\title{Helper object}
\format{\preformatted{Classes 'StatLogo', 'Stat', 'ggproto' <ggproto object: Class StatLogo, Stat>
    compute_group: function
    compute_layer: function
    compute_panel: function
    default_aes: uneval
    non_missing_aes: 
    parameters: function
    required_aes: x y
    retransform: TRUE
    setup_data: function
    setup_params: function
    super:  <ggproto object: Class Stat> 
}}
\usage{
StatLogo
}
\value{
proto object
}
\description{
shouldn't be used by the user, but has to be exported.
}
\keyword{datasets}

